Initial Commit
As big as this project is getting, it should DEFINITELY be versioned.
So, here we go. The initial checkin for my final project in my AI
course. At this point everything is incomplete and fairly undocumented.
The viewer has not been started yet (ugghh) and the runner is about
40% done. There is no README so figuring out what this program acutally
does may be a mystery (at the moment) for anyone who stumbles upon it.