Added code for Viewer
Well, the runner has been done for some time now, but
there has been no viewer available. Finally, we have a
viewer (web-based) that can be used to view the results.

It's a sinatra-based web-app that uses raphael to render
the graphs out from the result sets stored within the db.